fix: modularization of Details screen component
